requirements.txt
flask
flask-pymongo
flask-jwt-extended
flask-cors
qrcode
pillow
python-dotenv
gunicorn
dnspython
